+++ src/main/java/lumien/randomthings/container/redstoneinterface/ContainerAdvancedRedstoneInterface.java
@@ -10,16 +10,23 @@ import net.minecraft.entity.player.InventoryPlayer;
 import net.minecraft.inventory.Container;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
+import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
 public class ContainerAdvancedRedstoneInterface extends Container
 {
 	TileEntityAdvancedRedstoneInterface te;
+	BlockPos pos;
+	World world;
+	TileEntity expectedTileEntity;
 
 	public ContainerAdvancedRedstoneInterface(EntityPlayer player, World world, int x, int y, int z)
 	{
-		te = (TileEntityAdvancedRedstoneInterface) world.getTileEntity(new BlockPos(x, y, z));
+		this.world = world;
+		this.pos = new BlockPos(x, y, z);
+		te = (TileEntityAdvancedRedstoneInterface) world.getTileEntity(this.pos);
+		expectedTileEntity = te;
 
 		for (int i = 0; i < 9; i++)
 		{
@@ -99,6 +106,9 @@ public class ContainerAdvancedRedstoneInterface extends Container
 	@Override
 	public boolean canInteractWith(EntityPlayer playerIn)
 	{
-		return true;
+		return playerIn.getDistanceSq(this.pos.getX() + 0.5D, this.pos.getY() + 0.5D, this.pos.getZ() + 0.5D) <= 64.0D
+				&& expectedTileEntity != null
+				&& !expectedTileEntity.isInvalid()
+				&& this.world.getTileEntity(this.pos) == expectedTileEntity;
 	}
 }